requirements.txt
anthropic>=0.37.0
pinecone-client>=5.0.0
wikipedia>=1.4.0
sentence-transformers>=3.0.0
torch>=2.2.0
rank-bm25>=0.2.2
regex>=2024.4.16
tiktoken>=0.7.0
streamlit>=1.37.0
aiohttp>=3.9.5
httpx>=0.27.0
nest-asyncio>=1.6.0
cross-encoder>=0.1.5
